# spec/openapi.yaml
openapi: 3.0.3
info:
  title: HDGP Engine API
  description: |
    HDGP 人类尊严守护协议 Engine API。  
    Engine 为裁决组件，不接入 LLM；裁决逻辑由开发期写死的规则 + 治理流程后的规则包决定。
  version: 0.1.0-mvp
  contact:
    name: HDGP Protocol
    url: https://github.com/HumanDignityGuardian/HDGP-Core

servers:
  - url: http://localhost:8080
    description: 本地开发

paths:
  /hdgp/v1/evaluate:
    post:
      summary: 评估候选输出
      description: |
        对候选输出（如 LLM 回复）进行伦理/规则裁决。
        返回 verdict（allow/modify/block/fuse）、rules_triggered、actions。
      operationId: evaluate
      parameters:
        - name: include_report
          in: query
          required: false
          schema:
            type: string
            enum: ["true", "1", "false", "0"]
          description: 为 true 或 1 时，响应附带人可读的 report 字段
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateRequest"
      responses:
        "200":
          description: 评估结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateResponse"
        "400":
          description: 请求体无效（JSON 解析失败等）

  /hdgp/v1/status:
    get:
      summary: 健康检查与版本信息
      description: 返回 Engine 版本、spec 版本、policy 配置，用于可用性探测与版本校验。
      operationId: getStatus
      responses:
        "200":
          description: 状态信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"

components:
  schemas:
    EvaluateRequest:
      type: object
      required: [meta, subject, candidate]
      properties:
        meta:
          $ref: "#/components/schemas/Meta"
        subject:
          $ref: "#/components/schemas/Subject"
        input:
          $ref: "#/components/schemas/EvaluateInput"
        candidate:
          $ref: "#/components/schemas/Candidate"

    Meta:
      type: object
      properties:
        request_id: { type: string }
        scene:
          type: object
          properties:
            domain: { type: string }
            intent: { type: string }
            risk_level: { type: string }
            sensitivity: { type: array, items: { type: string } }
        policy:
          type: object
          properties:
            spec_version: { type: string }
            strategy_id: { type: string }
            bundles: { type: array, items: { type: string } }

    Subject:
      type: object
      required: [type]
      properties:
        type: { type: string }
        skill_id: { type: string }
        label: { type: string }

    EvaluateInput:
      type: object
      properties:
        prompt: { type: string }
        context: {}

    Candidate:
      type: object
      properties:
        text: { type: string }
        structured: {}
        raw: {}

    EvaluateResponse:
      type: object
      properties:
        request_id: { type: string }
        verdict: { type: string }
        actions: { type: array, items: { $ref: "#/components/schemas/Action" } }
        effective_output: { $ref: "#/components/schemas/EffectiveOutput" }
        rules_triggered: { type: array, items: { $ref: "#/components/schemas/RuleHit" } }
        engine_info: { $ref: "#/components/schemas/EngineInfo" }
        report: { type: string, description: 仅 include_report=true 时返回 }

    Action:
      type: object
      properties:
        type: { type: string }
        message: { type: string }
        details: { type: object }

    EffectiveOutput:
      type: object
      properties:
        text: { type: string }
        structured: {}

    RuleHit:
      type: object
      properties:
        rule_id: { type: string }
        principle_id: { type: string }
        article_id: { type: string }
        effect: { type: string }
        severity: { type: string }

    EngineInfo:
      type: object
      properties:
        version: { type: string }
        policy: { type: object }

    StatusResponse:
      type: object
      properties:
        engine_version: { type: string }
        spec_version: { type: string }
        policy:
          type: object
          properties:
            spec_version: { type: string }
            strategy_id: { type: string }
            bundles: { type: array, items: { type: string } }